move wgd depth detection to doubletime tools
